classify/requirements.txt
pytorch-cpu
torchvision-cpu
